requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.11.0
zope.event==6.0
zope.interface==8.0.1
